Removed interface IndicatorDAO
