+++ webcurator-core/src/main/java/org/webcurator/domain/IndicatorDAO.java
@@ -1,38 +1,110 @@
+/*
+ *  Copyright 2006 The National Library of New Zealand
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
 package org.webcurator.domain;
 
 import java.util.List;
 
+import org.apache.commons.logging.Log;
+import org.apache.commons.logging.LogFactory;
+import org.hibernate.query.Query;
+import org.hibernate.Session;
+import org.springframework.dao.DataAccessException;
+import org.springframework.orm.hibernate5.HibernateCallback;
+import org.springframework.orm.hibernate5.support.HibernateDaoSupport;
+import org.springframework.transaction.TransactionStatus;
+import org.springframework.transaction.annotation.Transactional;
+import org.springframework.transaction.support.TransactionCallback;
+import org.springframework.transaction.support.TransactionTemplate;
 import org.webcurator.domain.model.core.Indicator;
 
 /**
- * The object for accessing <code>Indicator</code>s from the persistent store.
+ * implements the IndicatorDAO Interface and provides the database calls for
+ * querying any objects related to indicator criteria. 
+ * @author oakleigh_sk
  */
-public interface IndicatorDAO {
+public class IndicatorDAO extends HibernateDaoSupport {
+    
+    private Log log = LogFactory.getLog(IndicatorDAO.class);
+    
+    private TransactionTemplate txTemplate = null;
 
-	/**
-	 * Save or update the specified object to the persistent data store.
-	 * @param aObject the object to save or update
-	 */
-    public void saveOrUpdate(Object aObject);
+    public void saveOrUpdate(final Object aObject) {
+        txTemplate.execute(
+                new TransactionCallback() {
+                    public Object doInTransaction(TransactionStatus ts) {
+                        try { 
+                            log.debug("Before Saving of Object");
+                            currentSession().saveOrUpdate(aObject);
+                            log.debug("After Saving Object");
+                        }
+                        catch(Exception ex) {
+                            log.warn("Setting Rollback Only",ex);
+                            ts.setRollbackOnly();
+                        }
+                        return null;
+                    }
+                }
+        );    
+    }
     
-	/**
-	 * Remove the specified object from the persistent data store.
-	 * @param aObject the object to remove
-	 */
-	void delete(Object aObject);
-	
-	 /**
-     * gets the specific <code>Indicator</code> based on the provided oid
-     * @param oid the <code>Indicator</code> Oid 
-     * @return the populated <code>Indicator</code> object
-     */
-    public Indicator getIndicatorByOid(Long oid);
- 
-    /**
-     * gets the <code>Indicator</code> for a selected <code>TargetInstance</code>
-     * @param agencyOid the Oid of the <code>TargetInstance</code> in which to search for <code>Indicator</code>s
-     * @return a List of fully populated <code>Indicator</code> objects
-     */
-    public List<Indicator> getIndicatorsByTargetInstanceOid(Long targetInstanceOid);
-    
+    public void delete(final Object aObject) {
+        txTemplate.execute(
+                new TransactionCallback() {
+                    public Object doInTransaction(TransactionStatus ts) {
+                        try {
+                            log.debug("Before Delete of Object");
+                            getHibernateTemplate().delete(aObject);
+                            log.debug("After Deletes Object");
+                        }
+                        catch (DataAccessException e) {
+                            log.warn("Setting Rollback Only",e);
+                            ts.setRollbackOnly();
+                            throw e;
+                        }
+                        return null;
+                    }
+                }
+        );    
+    }
+
+    @Transactional
+    public Indicator getIndicatorByOid(final Long indicatorOid) {
+        return (Indicator)getHibernateTemplate().execute(
+                new HibernateCallback() {
+                    public Object doInHibernate(Session session) {
+                        Query query = session.getNamedQuery(Indicator.QRY_GET_INDICATOR_BY_OID);
+                        query.setParameter(1, indicatorOid);
+                        return query.uniqueResult();
+                    }
+                }
+            );
+          
+    }
+
+    public List<Indicator> getIndicatorsByTargetInstanceOid(Long targetInstanceOid) {
+        List<Indicator> results = getHibernateTemplate().execute(session ->
+                session.getNamedQuery(Indicator.QRY_GET_INDICATORS_BY_TI_OID)
+                        .setParameter(1, targetInstanceOid)
+                        .list());
+
+        return results;
+    }
+
+    public void setTxTemplate(TransactionTemplate txTemplate) {
+        this.txTemplate = txTemplate;
+    }
+
 }
